New wet deposition
This update includes three major parts:

1) Wet deposition now follows the same scheme as in FLEXPART-9.02, i.e.
there is now a distinction between in-cloud and below-cloud scavenging
and no scavenging occurs above the highest cloud top. Clouds are
identified using CLDFRA 4D field from WRF.

2) Hourly precipitation and cloud cover calculated from WRF output.
In order to get the wet deposition working additional input to
flexpart-wrf was required. In previous versions precipitation and cloud
cover was set to zero, effectively turning off wet deposition all
together.  In this version, precipitation is read from the fields:
RAINNC,SNOWNC,HAILNC,GRAUPELNC ("large scale" i.e. grid scale)
RAINC,RAINSH ("convective", i.e. subgrid)
ALL OF THESE FIELDS MUST BE PRESENT IN THE OUTPUT DATA OR FLEXPART-WRF
WILL TERMINATE THE RUN! Default output settings in WRF3.4 should work
(just remember to add inverse density to record history)
These fields provide accumulated precipitation so the precipitation rate
is calculated based on these fields and the time span between input
files.

3) NetCDF output is now (partially) supported by setting IOUTTYPE=2
(in COMMAND).  The format mimics that of WRF and will hopefully work
with some WRF post-processing utilities. It also works well with ncview
for quick evaluations.  So far, concentration, wet- and dry deposition
is supported.  Mixing ratio is not supported, but the inclusion of this
is analogous to concentration in concoutput.f, the m-ratio field is being
initialized by writeheader.f but not no data is written. Backward runs
are not supported at all. Receptor points currently written to ascii but
could be included as additional time series.  Trajectories and particle
positions are not included in the netcdf, but written to ascii.  This
shoud not be written to netcdf format as it would probably be very space
inefficient, requiring fixed dimensions of size equal to the number of
particles released.

That said, I should also add that the wet deposition should undergo some
testing before used extensively.  I've sorted out obvious bugs but it
could still be misbehaving in more subtle ways.