Bug fix
There was a bug in writeheader*.f causing a crash when more than six
release points where assigned. It should now support as many as are
allowed by includepar (default is maxpoint=10000).

The same fix was also applied to writeheader_nests.f, however, please
note that nested output is not currently supported for netcdf output!